suma de la suma

+++ suma_de_la_suma_de_la_suma/src/main.rs
@@ -8,13 +8,13 @@ fn main() {
         .read_line(&mut n)
         .expect("Failed to read line");
 
-    let n: u32 = n.trim().parse().expect("Not a number");
+    let n: u128 = n.trim().parse().expect("Not a number");
 
-    let mut sum: u32 = 0;
-    let mut sum_total: u32 = 0;
+    let mut sum: u128 = 0;
+    let mut sum_total: u128 = 0;
 
     for i in 1..=n {
-        let aux_sum: u32 = i * (i + 1) / 2;
+        let aux_sum: u128 = i * (i + 1) / 2;
 
         sum_total += aux_sum + sum;
         sum += aux_sum;
